requirements.txt
Django==3.2.13
pytz==2021.1
sqlparse==0.4.1
psycopg2>=2.9.3
djangorestframework==3.13.1
djangorestframework-simplejwt==4.6.0
drf-yasg==1.20.0
django-filter==21.1
redis>=3.5
celery>=5.0
django-celery-beat==2.2.1
